+++ src/adapter/http/tests/tests/test_data_ingest.rs
@@ -608,6 +608,7 @@ impl DataIngestHarness {
             .add::<PushIngestPlannerImpl>()
             .add::<EngineProvisionerNull>()
             .add::<UploadServiceLocal>()
+            .add::<PushIngestDataUseCaseImpl>()
             .add_value(FileUploadLimitConfig::new_in_bytes(1000))
             .build();