+++ src/adapter/odata/tests/tests/test_handlers.rs
@@ -517,7 +517,7 @@ impl TestHarness {
 
         let ingest_result = self
             .push_ingest_executor
-            .ingest_from_url(target.clone(), ingest_plan, url, None)
+            .execute_ingest(target.clone(), ingest_plan, DataSource::Url(url), None)
             .await
             .unwrap();