+++ src/infra/core/src/use_cases/mod.rs
@@ -10,6 +10,7 @@
 mod compact_dataset_use_case_impl;
 mod pull_dataset_use_case_impl;
 mod push_dataset_use_case_impl;
+mod push_ingest_data_use_case_impl;
 mod reset_dataset_use_case_impl;
 mod set_watermark_use_case_impl;
 mod verify_dataset_use_case_impl;
@@ -17,6 +18,7 @@ mod verify_dataset_use_case_impl;
 pub use compact_dataset_use_case_impl::*;
 pub use pull_dataset_use_case_impl::*;
 pub use push_dataset_use_case_impl::*;
+pub use push_ingest_data_use_case_impl::*;
 pub use reset_dataset_use_case_impl::*;
 pub use set_watermark_use_case_impl::*;
 pub use verify_dataset_use_case_impl::*;