+++ src/infra/core/tests/tests/test_compaction_services_impl.rs
@@ -22,6 +22,7 @@ use kamu::testing::DatasetDataHelper;
 use kamu::*;
 use kamu_accounts::CurrentAccountSubject;
 use kamu_core::auth;
+use messaging_outbox::DummyOutboxImpl;
 use odf::dataset::testing::create_test_dataset_from_snapshot;
 use odf::metadata::testing::MetadataFactory;
 use s3_utils::S3Context;
@@ -1113,8 +1114,7 @@ struct CompactTestHarness {
     dataset_storage_unit_writer: Arc<dyn odf::DatasetStorageUnitWriter>,
     compaction_planner: Arc<dyn CompactionPlanner>,
     compaction_executor: Arc<dyn CompactionExecutor>,
-    push_ingest_planner: Arc<dyn PushIngestPlanner>,
-    push_ingest_executor: Arc<dyn PushIngestExecutor>,
+    ingest_data_use_case: Arc<dyn PushIngestDataUseCase>,
     transform_helper: TransformTestHelper,
     verification_svc: Arc<dyn VerificationService>,
     current_date_time: DateTime<Utc>,
@@ -1157,6 +1157,8 @@ impl CompactTestHarness {
             .add_value(EngineConfigDatafusionEmbeddedIngest::default())
             .add::<PushIngestExecutorImpl>()
             .add::<PushIngestPlannerImpl>()
+            .add::<PushIngestDataUseCaseImpl>()
+            .add::<DummyOutboxImpl>()
             .add::<TransformRequestPlannerImpl>()
             .add::<TransformElaborationServiceImpl>()
             .add::<TransformExecutorImpl>()
@@ -1176,8 +1178,7 @@ impl CompactTestHarness {
             dataset_storage_unit_writer: catalog.get_one().unwrap(),
             compaction_planner: catalog.get_one().unwrap(),
             compaction_executor: catalog.get_one().unwrap(),
-            push_ingest_planner: catalog.get_one().unwrap(),
-            push_ingest_executor: catalog.get_one().unwrap(),
+            ingest_data_use_case: catalog.get_one().unwrap(),
             verification_svc: catalog.get_one().unwrap(),
             transform_helper,
             current_date_time,
@@ -1215,6 +1216,8 @@ impl CompactTestHarness {
             .add_value(EngineConfigDatafusionEmbeddedIngest::default())
             .add::<PushIngestExecutorImpl>()
             .add::<PushIngestPlannerImpl>()
+            .add::<PushIngestDataUseCaseImpl>()
+            .add::<DummyOutboxImpl>()
             .add::<TransformRequestPlannerImpl>()
             .add::<TransformElaborationServiceImpl>()
             .add::<TransformExecutorImpl>()
@@ -1234,8 +1237,7 @@ impl CompactTestHarness {
             dataset_storage_unit_writer: catalog.get_one().unwrap(),
             compaction_planner: catalog.get_one().unwrap(),
             compaction_executor: catalog.get_one().unwrap(),
-            push_ingest_planner: catalog.get_one().unwrap(),
-            push_ingest_executor: catalog.get_one().unwrap(),
+            ingest_data_use_case: catalog.get_one().unwrap(),
             transform_helper,
             verification_svc: catalog.get_one().unwrap(),
             current_date_time,
@@ -1405,35 +1407,20 @@ impl CompactTestHarness {
     async fn ingest_data(&self, data_str: String, target: ResolvedDataset) {
         let data = std::io::Cursor::new(data_str);
 
-        let ingest_plan = self
-            .push_ingest_planner
-            .plan_ingest(target.clone(), None, PushIngestOpts::default())
+        self.ingest_data_use_case
+            .execute(
+                &target,
+                DataSource::Stream(Box::new(data)),
+                PushIngestDataUseCaseOptions {
+                    source_name: None,
+                    source_event_time: None,
+                    is_ingest_from_upload: false,
+                    media_type: None,
+                },
+                None,
+            )
             .await
             .unwrap();
-
-        let ingest_result = self
-            .push_ingest_executor
-            .ingest_from_stream(target.clone(), ingest_plan, Box::new(data), None)
-            .await
-            .unwrap();
-
-        if let PushIngestResult::Updated {
-            old_head, new_head, ..
-        } = &ingest_result
-        {
-            target
-                .as_metadata_chain()
-                .set_ref(
-                    &odf::BlockRef::Head,
-                    new_head,
-                    odf::dataset::SetRefOpts {
-                        validate_block_present: true,
-                        check_ref_is: Some(Some(old_head)),
-                    },
-                )
-                .await
-                .unwrap();
-        }
     }
 
     async fn commit_set_licence_block(&self, dataset_ref: &odf::DatasetRef, head: &odf::Multihash) {