+++ src/infra/core/tests/tests/test_transform_services_impl.rs
@@ -18,6 +18,7 @@ use kamu::domain::engine::*;
 use kamu::domain::*;
 use kamu::*;
 use kamu_accounts::CurrentAccountSubject;
+use messaging_outbox::DummyOutboxImpl;
 use odf::dataset::testing::create_test_dataset_from_snapshot;
 use odf::metadata::testing::MetadataFactory;
 use tempfile::TempDir;
@@ -38,8 +39,7 @@ struct TransformTestHarness {
     transform_executor: Arc<dyn TransformExecutor>,
     compaction_planner: Arc<dyn CompactionPlanner>,
     compaction_executor: Arc<dyn CompactionExecutor>,
-    push_ingest_planner: Arc<dyn PushIngestPlanner>,
-    push_ingest_executor: Arc<dyn PushIngestExecutor>,
+    ingest_data_use_case: Arc<dyn PushIngestDataUseCase>,
 }
 
 impl TransformTestHarness {
@@ -63,6 +63,7 @@ impl TransformTestHarness {
             .bind::<dyn odf::DatasetStorageUnitWriter, odf::dataset::DatasetStorageUnitLocalFs>()
             .add::<odf::dataset::DatasetDefaultLfsBuilder>()
             .bind::<dyn odf::dataset::DatasetLfsBuilder, odf::dataset::DatasetDefaultLfsBuilder>()
+            .add::<auth::AlwaysHappyDatasetActionAuthorizer>()
             .add::<SystemTimeSourceDefault>()
             .add::<ObjectStoreRegistryImpl>()
             .add::<ObjectStoreBuilderLocalFs>()
@@ -73,6 +74,8 @@ impl TransformTestHarness {
             .add_value(EngineConfigDatafusionEmbeddedIngest::default())
             .add::<PushIngestExecutorImpl>()
             .add::<PushIngestPlannerImpl>()
+            .add::<PushIngestDataUseCaseImpl>()
+            .add::<DummyOutboxImpl>()
             .add_value(engine_provisioner)
             .bind::<dyn EngineProvisioner, TEngineProvisioner>()
             .add::<TransformRequestPlannerImpl>()
@@ -89,8 +92,7 @@ impl TransformTestHarness {
             dataset_storage_unit_writer: catalog.get_one().unwrap(),
             compaction_planner: catalog.get_one().unwrap(),
             compaction_executor: catalog.get_one().unwrap(),
-            push_ingest_planner: catalog.get_one().unwrap(),
-            push_ingest_executor: catalog.get_one().unwrap(),
+            ingest_data_use_case: catalog.get_one().unwrap(),
             transform_request_planner: catalog.get_one().unwrap(),
             transform_elab_svc: catalog.get_one().unwrap(),
             transform_executor: catalog.get_one().unwrap(),
@@ -220,35 +222,20 @@ impl TransformTestHarness {
     async fn ingest_data(&self, data_str: String, target: ResolvedDataset) {
         let data = std::io::Cursor::new(data_str);
 
-        let ingest_plan = self
-            .push_ingest_planner
-            .plan_ingest(target.clone(), None, PushIngestOpts::default())
+        self.ingest_data_use_case
+            .execute(
+                &target,
+                DataSource::Stream(Box::new(data)),
+                PushIngestDataUseCaseOptions {
+                    source_name: None,
+                    source_event_time: None,
+                    is_ingest_from_upload: false,
+                    media_type: None,
+                },
+                None,
+            )
             .await
             .unwrap();
-
-        let ingest_result = self
-            .push_ingest_executor
-            .ingest_from_stream(target.clone(), ingest_plan, Box::new(data), None)
-            .await
-            .unwrap();
-
-        if let PushIngestResult::Updated {
-            old_head, new_head, ..
-        } = &ingest_result
-        {
-            target
-                .as_metadata_chain()
-                .set_ref(
-                    &odf::BlockRef::Head,
-                    new_head,
-                    odf::dataset::SetRefOpts {
-                        validate_block_present: true,
-                        check_ref_is: Some(Some(old_head)),
-                    },
-                )
-                .await
-                .unwrap();
-        }
     }
 
     async fn elaborate_transform(