+++ src/infra/core/tests/tests/use_cases/mod.rs
@@ -10,6 +10,7 @@
 mod test_compact_dataset_use_case;
 mod test_pull_dataset_use_case;
 mod test_push_dataset_use_case;
+mod test_push_ingest_data_use_case;
 mod test_reset_dataset_use_case;
 mod test_set_watermark_use_case;
 mod test_verify_dataset_use_case;