+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Recommendation: for ease of reading, use the following order:
 - New `engine.datafusionEmbedded` config section allows to pass custom DataFusion settings when engine is used in igest, batch query, and compaction contexts.
 ### Fixed
 - Multiple performance improvements in batch queries to avoid unnecessary metadata scanning.
+- New `PushIngestDataUseCase` and used it in Http `/ingest` handler and `ingest_command`
 
 ## [0.231.0] - 2025-03-31
 ### Added